Add TestRunThreadContext and move ITestRun to the same package.
