+++ sdk/src/main/java/com/microsoft/hydralab/agent/runner/ITestRun.java
@@ -1,4 +1,4 @@
-package com.microsoft.hydralab;
+package com.microsoft.hydralab.agent.runner;
 
 import java.io.File;
